requirements.txt
langbot-plugin
aiosqlite
Pillow
